+++ backend/src/main/java/org/maklashev/goldengroup/model/repositories/TypesRepository.java
@@ -3,6 +3,6 @@ package org.maklashev.goldengroup.model.repositories;
 import org.maklashev.goldengroup.model.entity.Types;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypesRepository extends JpaRepository <Types, Long>{
+public interface TypesRepository extends JpaRepository <Types, Integer>{
 
 }
